pyproject.toml
[project]
name = "posefx-studio"
version = "0.1.0"
description = "Real-time webcam pipeline for body tracking, gesture detection, and live visual effects"
requires-python = ">=3.11"
dependencies = [
    "opencv-python>=4.8",
    "ultralytics>=8.1",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "pyaudio>=0.2.14",
    "scipy>=1.11",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
